move broadcast_tensor_dict into coordinator
